Cleaned up invoice generation in tests
no issue
