+++ packages/mg-stripe-csv/src/test/importing.test.ts
@@ -5,7 +5,7 @@ import {createCouponImporter} from '../lib/importers/createCouponImporter.js';
 import {createPriceImporter} from '../lib/importers/createPriceImporter.js';
 import {createProductImporter} from '../lib/importers/createProductImporter.js';
 import {createSubscriptionImporter} from '../lib/importers/createSubscriptionImporter.js';
-import {buildPrice, buildProduct, buildSubscription, createDeclinedCustomer, createValidCustomer, getStripeTestAPIKey} from './utils/stripe.js';
+import {buildInvoice, buildPrice, buildProduct, buildSubscription, createDeclinedCustomer, createValidCustomer, getStripeTestAPIKey} from './utils/stripe.js';
 import {Options} from '../lib/Options.js';
 import assert from 'assert/strict';
 import sinon from 'sinon';
@@ -238,124 +238,6 @@ describe('test', () => {
         });
         const now = Math.floor(new Date().getTime() / 1000);
 
-        // The whole point of Past Due invoices is that they have an open invoice that is not yet paid
-        // So to simulate this, we need to also generate and pass in the open invoice
-        const invoiceItem: Stripe.InvoiceLineItem = {
-            id: 'ii_1',
-            amount: 1000,
-            object: 'line_item',
-            amount_excluding_tax: null,
-            currency: 'usd',
-            description: null,
-            discount_amounts: null,
-            discountable: false,
-            discounts: null,
-            livemode: false,
-            metadata: {},
-            period: {
-                start: now - 15 * 24 * 60 * 60,
-                end: now + 15 * 24 * 60 * 60
-            },
-            plan: null,
-            price: oldPrice,
-            proration: false,
-            proration_details: null,
-            quantity: 1,
-            subscription: null,
-            type: 'subscription',
-            unit_amount_excluding_tax: null
-        };
-
-        const invoice: Stripe.Invoice = {
-            id: 'in_1',
-            object: 'invoice',
-            account_country: null,
-            account_name: null,
-            account_tax_ids: null,
-            amount_due: 0,
-            amount_paid: 0,
-            amount_remaining: 0,
-            amount_shipping: 0,
-            application: null,
-            application_fee_amount: null,
-            attempt_count: 0,
-            attempted: false,
-            automatic_tax: {
-                enabled: false,
-                status: null
-            },
-            billing_reason: null,
-            charge: null,
-            collection_method: 'charge_automatically',
-            created: 0,
-            currency: '',
-            custom_fields: null,
-            customer: null,
-            customer_address: null,
-            customer_email: null,
-            customer_name: null,
-            customer_phone: null,
-            customer_shipping: null,
-            customer_tax_exempt: null,
-            default_payment_method: null,
-            default_source: null,
-            default_tax_rates: [],
-            description: null,
-            discount: null,
-            discounts: null,
-            due_date: null,
-            effective_at: null,
-            ending_balance: null,
-            footer: null,
-            from_invoice: null,
-            last_finalization_error: null,
-            latest_revision: null,
-            lines: {
-                object: 'list',
-                data: [
-                    invoiceItem
-                ],
-                has_more: false,
-                url: ''
-            },
-            livemode: false,
-            metadata: null,
-            next_payment_attempt: null,
-            number: null,
-            on_behalf_of: null,
-            paid: false,
-            paid_out_of_band: false,
-            payment_intent: null,
-            payment_settings: {} as any,
-            period_end: 0,
-            period_start: 0,
-            post_payment_credit_notes_amount: 0,
-            pre_payment_credit_notes_amount: 0,
-            quote: null,
-            receipt_number: null,
-            rendering_options: null,
-            shipping_cost: null,
-            shipping_details: null,
-            starting_balance: 0,
-            statement_descriptor: null,
-            status: 'open',
-            status_transitions: {} as any,
-            subscription: null,
-            subtotal: 0,
-            subtotal_excluding_tax: null,
-            tax: null,
-            test_clock: null,
-            total: 0,
-            total_discount_amounts: null,
-            total_excluding_tax: null,
-            total_tax_amounts: [],
-            transfer_data: null,
-            webhooks_delivered_at: null
-        };
-
-        // Return these invoices when we ask Stripe
-        currentInvoices = [invoice];
-
         const oldSubscription: Stripe.Subscription = buildSubscription({
             status: 'past_due',
             customer: declinedCustomer.id,
@@ -365,6 +247,23 @@ describe('test', () => {
                 }
             ]
         });
+
+        const invoice: Stripe.Invoice = buildInvoice({
+            customer: declinedCustomer.id,
+            subscription: oldSubscription.id,
+            lines: [
+                {
+                    price: oldPrice,
+                    period: {
+                        start: now - 15 * 24 * 60 * 60,
+                        end: now + 15 * 24 * 60 * 60
+                    }
+                }
+            ]
+        });
+
+        // Return these invoices when we ask Stripe
+        currentInvoices = [invoice];
 
         const newSubscriptionId = await subscriptionImporter.recreate(oldSubscription);
         const newSubscription = await stripe.client.subscriptions.retrieve(newSubscriptionId);
@@ -390,7 +289,15 @@ describe('test', () => {
         assert.equal(newInvoices.data.length, 1);
         assert.equal(newInvoices.data[0].status, 'open');
         assert.equal(newInvoices.data[0].attempt_count, 1);
-        assert.equal(newInvoices.data[0].amount_due, 1000);
+        assert.equal(newInvoices.data[0].amount_due, 100);
         assert.equal(newInvoices.data[0].customer, declinedCustomer.id);
+
+        // Check upcoming invoice
+        const upcomingInvoice = await stripe.client.invoices.retrieveUpcoming({
+            customer: declinedCustomer.id,
+            subscription: newSubscription.id
+        });
+        assert.equal(upcomingInvoice.amount_due, 100);
+        assert.equal(upcomingInvoice.lines.data[0].period.start, oldSubscription.current_period_end);
     });
 });
